修改tinyint 为Integer类型，以前byte类型无法转换 修改时间转换值可能是字符串bug 添加表的前缀后缀表达式解析 添加cdcSource额外添加一个参数sinkTimeColumn（数据入湖时间）

+++ dlink-metadata/dlink-metadata-mysql/src/main/java/com/dlink/metadata/convert/MySqlTypeConvert.java
@@ -19,7 +19,7 @@ public class MySqlTypeConvert implements ITypeConvert {
         }
         String t = column.getType().toLowerCase();
         if (t.contains("tinyint")) {
-            columnType = ColumnType.BYTE;
+            columnType = ColumnType.INTEGER;
         } else if (t.contains("smallint") || t.contains("tinyint unsigned")) {
             columnType = ColumnType.SHORT;
         } else if (t.contains("bigint unsigned") || t.contains("numeric") || t.contains("decimal")) {
